refactor: remove uploadSetupMessage and related types, introducing new setupMessage module for improved organization and clarity
